Integration tests passing
