+++ tests/services/contracts/IDContract.test.ts
@@ -1,5 +1,5 @@
 import { IDContract, sha256 } from 'tonomy-id-sdk';
-import { createRandomID } from '../../user.test';
+import { createRandomID } from '../../util/user';
 
 const idContract = IDContract.Instance;
 
@@ -17,8 +17,7 @@ describe("IDContract tests", () => {
         expect(idInfo.username_hash).toEqual(sha256(user.username));
         expect(idInfo.status).toEqual(0); // 0 = Creating
         expect(idInfo.type).toEqual(0); // 0 = Person
-        // expect(idInfo.password_salt).toEqual(user.salt); // TODO uncomment when salt implemented properly in User.ts
-
+        expect(idInfo.password_salt).toEqual(user.salt.toString());
 
         // get by username
         idInfo = await idContract.getAccountTonomyIDInfo(user.username);
